requirements.txt
python-dotenv
pyautogui
pyperclip
opencv-python
pillow